adds secondary menu items to about tab
