+++ cypress/integration/mobile/visitorCanNavigateUsingHeader.feature.js
@@ -26,20 +26,16 @@ describe("visitor can navigate between views", () => {
 
     it("is expected to navigate to about organization page", () => {
       cy.get("[data-cy=about-tab]").click();
+      cy.get("[data-cy=organization-tab]").should("contain", "organization")
+      cy.get("[data-cy=self-care-tab]").should("contain", "self care")
+      cy.get("[data-cy=organization-tab]").click()
       cy.url().should("include", "/about/organization");
     });
 
     it("is expected to navigate to about self care page", () => {
       cy.get("[data-cy=about-tab]").click();
-      cy.get("[data-cy=self-care-sub-tab]").click();
+      cy.get("[data-cy=self-care-tab]").click();
       cy.url().should("include", "/about/self_care");
-    });
-
-    it("is expected to navigate back to about organization page", () => {
-      cy.get("[data-cy=about-tab]").click();
-      cy.get("[data-cy=self-care-sub-tab]").click();
-      cy.get("[data-cy=organization-sub-tab]").click();
-      cy.url().should("include", "/about/organization");
     });
 
     it("is expected to navigate to about home page", () => {
